Collapse rpcproxy_internal_test.go into rpcproxy_test.go
Requested by: @dadgar
